Created Food Fields Viewable In Console
- Once Add Button Is Pressed, The Console Will Display All Data In Food
Entry Fields
- Alert View Now Displayed On Pressing Add Button If Name Field Empty
- Minor Fixes

+++ My Fit Diet/FoodObject.h
@@ -16,10 +16,10 @@
 @property float servingSize;
 
 @property int calories;
-@property int totalFats;
-@property int saturatedFats;
-@property int sodium;
-@property int totalCarbohydrates;
-@property int protein;
+@property float totalFats;
+@property float saturatedFats;
+@property float sodium;
+@property float totalCarbohydrates;
+@property float protein;
 
 @end
